conversational_agent/requirements.txt
fastapi
uvicorn
google-cloud
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-firestore
pydantic
chainlit
httpx
